% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afghan.R
\docType{data}
\name{afghan}
\alias{afghan}
\title{Afghanistan Survey Data}
\format{A data frame with 2754 rows and 11 variables:
\describe{
 \item{ province }{ character }  \item{ district }{ character }  \item{ village.id }{ integer }  \item{ age }{ integer }  \item{ educ.years }{ integer }  \item{ employed }{ integer }  \item{ income }{ character }  \item{ violent.exp.ISAF }{ integer }  \item{ violent.exp.taliban }{ integer }  \item{ list.group }{ character }  \item{ list.response }{ integer } 
}}
\usage{
afghan
}
\description{
Afghanistan Survey Data
}
\details{
See \emph{QSS} Table 3.1.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
