% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afghan.village.R
\docType{data}
\name{afghan.village}
\alias{afghan.village}
\title{Afghanistan Village Data}
\format{A data frame with 1864 rows and 3 variables:
\describe{
 \item{ altitude }{ numeric }  \item{ population }{ integer }  \item{ village.surveyed }{ integer } 
}}
\usage{
afghan.village
}
\description{
Afghanistan Village Data
}
\details{
See \emph{QSS} Table 3.2.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
