% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canada2011.R
\docType{data}
\name{canada2011}
\alias{canada2011}
\title{Canada Election Data (2011)}
\format{A data frame with 66060 rows and 3 variables:
\describe{
 \item{ votes }{ integer }  \item{ turnout }{ integer }  \item{ N }{ integer } 
}}
\usage{
canada2011
}
\description{
Canada Election Data (2011)
}
\details{
See \emph{QSS} Table 6.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

