% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccap2012.R
\docType{data}
\name{ccap2012}
\alias{ccap2012}
\title{Cooperative Campaign Analysis Project (CCAP) Data}
\format{A data frame with 43998 rows and 3 variables:
\describe{
 \item{ .row }{ integer }  \item{ caseid }{ integer }  \item{ gaytherm }{ integer } 
}}
\usage{
ccap2012
}
\description{
Cooperative Campaign Analysis Project (CCAP) Data
}
\details{
See \emph{QSS} Table 3.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

