% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chechen.R
\docType{data}
\name{chechen}
\alias{chechen}
\title{Insurgent attacks and Russian shelling in Chechnya}
\format{A data frame with 318 rows and 6 variables:
\describe{
 \item{ village }{ character }  \item{ groznyy }{ integer }  \item{ fire }{ integer }  \item{ deaths }{ integer }  \item{ preattack }{ integer }  \item{ postattack }{ integer } 
}}
\usage{
chechen
}
\description{
Insurgent attacks and Russian shelling in Chechnya
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

