% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinawomen.R
\docType{data}
\name{chinawomen}
\alias{chinawomen}
\title{Sex Ratio and the Price of Agricultural Crops in China}
\format{A data frame with 51766 rows and 9 variables:
\describe{
 \item{ admin }{ integer }  \item{ biryr }{ integer }  \item{ birpop }{ integer }  \item{ han }{ numeric }  \item{ sex }{ numeric }  \item{ teasown }{ numeric }  \item{ orch }{ numeric }  \item{ cashcrop }{ numeric }  \item{ post }{ integer } 
}}
\usage{
chinawomen
}
\description{
Sex Ratio and the Price of Agricultural Crops in China
}
\details{
See \emph{QSS} Table 7.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

