% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constitution.R
\docType{data}
\name{constitution}
\alias{constitution}
\title{Analyzing the Preambles of Constitutions}
\format{A data frame with 155 rows and 3 variables:
\describe{
 \item{ country }{ character: country name with words separated by underscores }
 \item{ year }{ integer: year the constitution was created }
 \item{ preamble }{ character: raw text of the preamble to the constitution }
}}
\usage{
constitution
}
\description{
Contains Preamble text for 155 countries.
}
\details{
See \emph{QSS} Table 5.6.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { David S. Law and Mila Versteeg. (2012). 'The Declining Influence of the United States
Constitution', \emph{New York University Law Review} Vol. 87, No. 3, pp. 762--858,
and Zachary Elkins, Tom Ginsburg. }
\item { James Melton. (2012). 'Comments on Law And Versteeg's the Declining Influence of the United
States Constitution.' \emph{New York University Law Review} Vol. 87, No. 6, pp. 2088--2101 }
}
}
\keyword{datasets}
