% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\docType{data}
\name{elections}
\alias{elections}
\title{County-Level US Presidential Election Data}
\format{A data frame with 43494 rows and 6 variables:
\describe{
 \item{ year }{ integer }  \item{ state }{ character }  \item{ county }{ character }  \item{ rep }{ integer }  \item{ dem }{ integer }  \item{ other }{ integer } 
}}
\usage{
elections
}
\description{
County-Level US Presidential Election Data
}
\details{
See \emph{QSS} Table 5.8.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

