% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face.R
\docType{data}
\name{face}
\alias{face}
\title{Facial Appearanace Experiment Data}
\format{A data frame with 119 rows and 10 variables:
\describe{
 \item{ year }{ integer }  \item{ state }{ character }  \item{ winner }{ character }  \item{ loser }{ character }  \item{ w.party }{ character }  \item{ l.party }{ character }  \item{ d.comp }{ numeric }  \item{ r.comp }{ numeric }  \item{ d.votes }{ integer }  \item{ r.votes }{ integer } 
}}
\usage{
face
}
\description{
Facial Appearanace Experiment Data
}
\details{
See \emph{QSS} Table 4.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
