% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedrawer.R
\docType{data}
\name{filedrawer}
\alias{filedrawer}
\title{TESS Program Academic Research Data}
\format{A data frame with 221 rows and 5 variables:
\describe{
 \item{ id }{ integer: study identifier }
 \item{ journal }{ character: discipline of the journal for published articles }
 \item{ DV }{ character: publication status }
 \item{ IV }{ character: statistical significance of the main findings }
 \item{ max.h }{ integer: H-index (highest among authors) }
}}
\usage{
filedrawer
}
\description{
Contains information about 221 research projects funded by the TESS program
(Time-Sharing Experiments in the Social Sciences).
}
\details{
See \emph{QSS} Table 7.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
