% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida.R
\docType{data}
\name{florida}
\alias{florida}
\title{1996 and 2000 US Presidential Election Data for Florida Counties}
\format{A data frame with 67 rows and 7 variables:
\describe{
 \item{ county }{ character }  \item{ Clinton96 }{ integer }  \item{ Dole96 }{ integer }  \item{ Perot96 }{ integer }  \item{ Bush00 }{ integer }  \item{ Gore00 }{ integer }  \item{ Buchanan00 }{ integer } 
}}
\usage{
florida
}
\description{
1996 and 2000 US Presidential Election Data for Florida Counties
}
\details{
See \emph{QSS} Table 4.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

