% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVoters.R
\docType{data}
\name{FLVoters}
\alias{FLVoters}
\title{Florida Registered Voter List Sample Data}
\format{A data frame with 10000 rows and 6 variables:
\describe{
 \item{ surname }{ character }  \item{ county }{ integer }  \item{ VTD }{ integer }  \item{ age }{ integer }  \item{ gender }{ character }  \item{ race }{ character } 
}}
\usage{
FLVoters
}
\description{
Florida Registered Voter List Sample Data
}
\details{
See \emph{QSS} Table 6.1.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
