% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gay.R
\docType{data}
\name{gay}
\alias{gay}
\title{Gay Marriage Study Data}
\format{A data frame with 69592 rows and 4 variables:
\describe{
 \item{ study }{ integer: source of the data (1 = study 1, 2 = study 2) }
 \item{ treatment }{ character: five possible treatment assignment options }
 \item{ wave }{ integer: survey wave (a total of seven waves)}
 \item{ ssm }{ integer: five-point scale on same-sex marriage, higher scores indicate support }
}}
\usage{
gay
}
\description{
Data from an experiment that analyzed whether contact with a gay canvasser affected
opinions on gay marriage.
}
\details{
See \emph{QSS} Table 2.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
