% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrade08.R
\docType{data}
\name{intrade08}
\alias{intrade08}
\title{Intrade Betting Market Data}
\format{A data frame with 36891 rows and 10 variables:
\describe{
 \item{ .row }{ integer }  \item{ day }{ Date }  \item{ statename }{ character }  \item{ MarketD }{ character }  \item{ PriceD }{ numeric }  \item{ VolumeD }{ integer }  \item{ MarketR }{ character }  \item{ PriceR }{ numeric }  \item{ VolumeR }{ integer }  \item{ state }{ character } 
}}
\usage{
intrade08
}
\description{
Intrade Betting Market Data
}
\details{
See \emph{QSS} Table 4.9.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

