% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intrade12.R
\docType{data}
\name{intrade12}
\alias{intrade12}
\title{Intrade Prediction Market Data}
\format{A data frame with 32623 rows and 8 variables:
\describe{
 \item{ .row }{ integer }  \item{ day }{ Date }  \item{ statename }{ character }  \item{ PriceD }{ numeric }  \item{ VolumeD }{ integer }  \item{ PriceR }{ numeric }  \item{ VolumeR }{ integer }  \item{ state }{ character } 
}}
\usage{
intrade12
}
\description{
Intrade Prediction Market Data
}
\details{
See \emph{QSS} Table 4.9.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
