% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kenya.R
\docType{data}
\name{Kenya}
\alias{Kenya}
\title{Fertility and Mortality estimates data (Kenya)}
\format{A data frame with 30 rows and 8 variables:
\describe{
 \item{ country }{ character }  \item{ period }{ character }  \item{ age }{ character }  \item{ births }{ numeric }  \item{ deaths }{ numeric }  \item{ py.men }{ numeric }  \item{ py.women }{ numeric }  \item{ l_x }{ numeric } 
}}
\usage{
Kenya
}
\description{
Fertility and Mortality estimates data (Kenya)
}
\details{
See \emph{QSS} Table 1.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

