% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minwage.R
\docType{data}
\name{minwage}
\alias{minwage}
\title{Minimum Wage Study Data}
\format{A data frame with 358 rows and 8 variables:
\describe{
 \item{ chain }{ character }  \item{ location }{ character }  \item{ wageBefore }{ numeric }  \item{ wageAfter }{ numeric }  \item{ fullBefore }{ numeric }  \item{ fullAfter }{ numeric }  \item{ partBefore }{ numeric }  \item{ partAfter }{ numeric } 
}}
\usage{
minwage
}
\description{
Minimum Wage Study Data
}
\details{
See \emph{QSS} Table 2.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

