% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nazis.R
\docType{data}
\name{nazis}
\alias{nazis}
\title{Analysis of the 1932 German Election}
\format{A data frame with 681 rows and 7 variables:
\describe{
 \item{ shareself }{ numeric }  \item{ shareblue }{ numeric }  \item{ sharewhite }{ numeric }  \item{ sharedomestic }{ numeric }  \item{ shareunemployed }{ numeric }  \item{ nvoter }{ integer }  \item{ nazivote }{ integer } 
}}
\usage{
nazis
}
\description{
Analysis of the 1932 German Election
}
\details{
See \emph{QSS} Table 7.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
