% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls12.R
\docType{data}
\name{polls12}
\alias{polls12}
\title{2012 US Presidential Polls Data}
\format{A data frame with 1227 rows and 6 variables:
\describe{
 \item{ .row }{ integer }  \item{ state }{ character }  \item{ Obama }{ integer }  \item{ Romney }{ integer }  \item{ Pollster }{ character }  \item{ middate }{ Date } 
}}
\usage{
polls12
}
\description{
2012 US Presidential Polls Data
}
\details{
See \emph{QSS} Table 4.2.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
