% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollsUS08.R
\docType{data}
\name{pollsUS08}
\alias{pollsUS08}
\title{2008 US Presidential National Polls Data}
\format{A data frame with 526 rows and 4 variables:
\describe{
 \item{ Pollster }{ character: name of the organization conducting the poll }
 \item{ McCain }{ integer: predicted support for McCain (percentage) }
 \item{ Obama }{ integer: predicted support for Obama (percentage) }
 \item{ middate }{ Date: middate of the period when the poll was conducted }
}}
\usage{
pollsUS08
}
\description{
2008 US Presidential National Polls Data
}
\details{
See \emph{QSS} Table 4.2.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
