% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres08.R
\docType{data}
\name{pres08}
\alias{pres08}
\title{2008 US Presidential Election Data}
\format{A data frame with 51 rows and 5 variables:
\describe{
 \item{ state.name }{ character }  \item{ state }{ character }  \item{ Obama }{ integer }  \item{ McCain }{ integer }  \item{ EV }{ integer } 
}}
\usage{
pres08
}
\description{
2008 US Presidential Election Data
}
\details{
See \emph{QSS} Table 4.1.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

