% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres12.R
\docType{data}
\name{pres12}
\alias{pres12}
\title{2012 US Presidential Election Data}
\format{A data frame with 51 rows and 4 variables:
\describe{
 \item{ state }{ character }  \item{ Obama }{ integer }  \item{ Romney }{ integer }  \item{ EV }{ integer } 
}}
\usage{
pres12
}
\description{
2012 US Presidential Election Data
}
\details{
See \emph{QSS} Table 4.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
