% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/published.R
\docType{data}
\name{published}
\alias{published}
\title{Published Journal Article Data}
\format{A data frame with 53 rows and 5 variables:
\describe{
 \item{ id.p }{ integer: paper identifier }
 \item{ cond.s }{ integer: number of conditions in the study }
 \item{ out.s }{ integer: number of outcome variables in the study }
 \item{ cond.p }{ integer: number of conditions presented in the paper }
 \item{ out.p }{ integer: number of outcome variables used in the paper }
}}
\usage{
published
}
\description{
Published Journal Article Data
}
\details{
See \emph{QSS} Table 7.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
