% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\docType{data}
\name{resources}
\alias{resources}
\title{Oil, Democracy, and Development}
\format{A data frame with 4530 rows and 8 variables:
\describe{
 \item{ cty_name }{ character }  \item{ year }{ integer }  \item{ logGDPcp }{ numeric }  \item{ regime }{ numeric }  \item{ oil }{ numeric }  \item{ metal }{ numeric }  \item{ illit }{ numeric }  \item{ life }{ numeric } 
}}
\usage{
resources
}
\description{
Oil, Democracy, and Development
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
