% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume.R
\docType{data}
\name{resume}
\alias{resume}
\title{Resume Experiment Data}
\format{A data frame with 4870 rows and 4 variables:
\describe{
 \item{ firstname }{ character }  \item{ sex }{ character }  \item{ race }{ character }  \item{ call }{ integer } 
}}
\usage{
resume
}
\description{
Resume Experiment Data
}
\details{
See \emph{QSS} Table 2.1.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
