% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia2003.R
\docType{data}
\name{russia2003}
\alias{russia2003}
\title{Russian Election Data (2003)}
\format{A data frame with 94972 rows and 3 variables:
\describe{
 \item{ N }{ integer: total number of voters in a precinct }
 \item{ turnout }{ integer: total turnout in a precinct }
 \item{ votes }{ integer: total number of votes for the winner in a precinct }
}}
\usage{
russia2003
}
\description{
Russian Election Data (2003)
}
\details{
See \emph{QSS} Table 6.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
