% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia2012.R
\docType{data}
\name{russia2012}
\alias{russia2012}
\title{Russian Election Data (2012)}
\format{A data frame with 95552 rows and 3 variables:
\describe{
 \item{ N }{ integer }  \item{ turnout }{ integer }  \item{ votes }{ integer } 
}}
\usage{
russia2012
}
\description{
Russian Election Data (2012)
}
\details{
See \emph{QSS} Table 6.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

