% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\docType{data}
\name{STAR}
\alias{STAR}
\title{STAR Project Data}
\format{A data frame with 6325 rows and 6 variables:
\describe{
 \item{ race }{ integer: student’s race (white = 1, black = 2, Asian = 3, Hispanic = 4, Native American = 5, others = 6) }
 \item{ classtype }{ integer: type of kindergarten class (small = 1, regular = 2, regular with aid = 3) }
 \item{ yearssmall }{ integer: number of years in small classes }
 \item{ hsgrad }{ integer: high-school graduation (did graduate = 1, did not graduate = 0) }
 \item{ g4math }{ integer: total scaled score for the math portion of the fourth-grade standardized test }
 \item{ g4reading }{ integer: total scaled score for the reading portion of the fourth-grade standardized test }
}}
\usage{
STAR
}
\description{
Contains data from the STAR (Student–Teacher Achievement Ratio) Project,
a four-year longitudinal study examining the effect of class size in early grade levels on educational
performance and personal development
}
\details{
See \emph{QSS} Table 2.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
