% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\docType{data}
\name{trade}
\alias{trade}
\title{International Trade Network Data}
\format{A data frame with 114980 rows and 5 variables:
\describe{
 \item{ .row }{ integer: row number }
 \item{ country1 }{ character: country name of exporter }
 \item{ country2 }{ character: country name of importer }
 \item{ year }{ integer: year }
 \item{ exports }{ numeric: total value of exports (in tens of millions of dollars) }
}}
\usage{
trade
}
\description{
Contains international trade data; value of exports from one country to another in a given year.
}
\details{
See \emph{QSS} Table 5.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
