% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\docType{data}
\name{transfer}
\alias{transfer}
\title{Brazilian Government Transfer Data}
\format{A data frame with 1787 rows and 9 variables:
\describe{
 \item{ .row }{ integer: row number }
 \item{ state }{ character: state name }
 \item{ region }{ character: region name }
 \item{ literate91 }{ numeric: literacy rate of the state in 1991 }
 \item{ educ91 }{ numeric: average years in education of the state in 1991 }
 \item{ poverty91 }{ numeric: poverty rate of the state in 1991 }
 \item{ poverty80 }{ numeric: poverty rate of the state in 1980 }
 \item{ educ80 }{ numeric: average years in education of the state in 1980 }
 \item{ pop82 }{ integer: population in 1982 }
}}
\usage{
transfer
}
\description{
Data set used in experiment to estimate the
effects of increased Brazilian government spending on educational
attainment, literacy, and poverty rates.
}
\details{
See \emph{QSS} Table 4.12.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { Stephan Litschig and Kevin M. Morrison (2013) “The impact of intergovernmental
transfers on education outcomes and poverty reduction.” \emph{American Economic Journal: Applied Economics},
vol. 5, no. 4, pp. 206–240. doi = 10.1257/app.5.4.206 }
}
}
\keyword{datasets}
