% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnout.R
\docType{data}
\name{turnout}
\alias{turnout}
\title{US election turnout data}
\format{A data frame with 14 rows and 9 variables:
\describe{
 \item{ year }{ integer }  \item{ VEP }{ integer }  \item{ VAP }{ integer }  \item{ total }{ integer }  \item{ ANES }{ integer }  \item{ felons }{ integer }  \item{ noncit }{ integer }  \item{ overseas }{ integer }  \item{ osvoters }{ integer } 
}}
\usage{
turnout
}
\description{
US election turnout data
}
\details{
See \emph{QSS} Table 1.3.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
