% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.following.R
\docType{data}
\name{twitter.following}
\alias{twitter.following}
\title{Senator Follow / Following Data}
\format{A data frame with 3859 rows and 2 variables:
\describe{
 \item{ following }{ character }  \item{ followed }{ character } 
}}
\usage{
twitter.following
}
\description{
Senator Follow / Following Data
}
\details{
See \emph{QSS} Table 5.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

