% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USGini.R
\docType{data}
\name{USGini}
\alias{USGini}
\title{US Gini Coefficient Data}
\format{A data frame with 67 rows and 3 variables:
\describe{
 \item{ .row }{ integer: row number }
 \item{ year }{ integer: year }
 \item{ gini }{ numeric: US Gini coefficient }
}}
\usage{
USGini
}
\description{
Contains the Gini coefficient for the United States from 1947 to 2013.
}
\details{
See \emph{QSS} Table 3.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
