% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\docType{data}
\name{vignettes}
\alias{vignettes}
\title{Political Efficacy Data}
\format{A data frame with 781 rows and 6 variables:
\describe{
 \item{ self }{ integer }  \item{ alison }{ integer }  \item{ jane }{ integer }  \item{ moses }{ integer }  \item{ china }{ integer }  \item{ age }{ integer } 
}}
\usage{
vignettes
}
\description{
Political Efficacy Data
}
\details{
See \emph{QSS} Table 3.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
