% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walmart.R
\docType{data}
\name{walmart}
\alias{walmart}
\title{Walmart Store Opening Data}
\format{A data frame with 3251 rows and 7 variables:
\describe{
 \item{ opendate }{ Date }  \item{ st.address }{ character }  \item{ city }{ character }  \item{ state }{ character }  \item{ long }{ numeric }  \item{ lat }{ numeric }  \item{ type }{ character } 
}}
\usage{
walmart
}
\description{
Walmart Store Opening Data
}
\details{
See \emph{QSS} Table 5.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
