% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/women.R
\docType{data}
\name{women}
\alias{women}
\title{Women as Policy Makers Experiment}
\format{A data frame with 322 rows and 6 variables:
\describe{
 \item{ GP }{ integer }  \item{ village }{ integer }  \item{ reserved }{ integer }  \item{ female }{ integer }  \item{ irrigation }{ integer }  \item{ water }{ integer } 
}}
\usage{
women
}
\description{
Women as Policy Makers Experiment
}
\details{
See \emph{QSS} Table 4.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

