% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccap2012.R
\docType{data}
\name{ccap2012}
\alias{ccap2012}
\title{Cooperative Campaign Analysis Project (CCAP) Data}
\format{A data frame with 43998 rows and 3 variables:
\describe{
 \item{ .row }{ integer: row number }
 \item{ caseid }{ integer: unique respondent ID }
 \item{ gaytherm }{ integer: survey thermometer rating of feeling towards gay couples (0–100) }
}}
\usage{
ccap2012
}
\description{
CCAP data set alleged to have been used as the basis for the gay marriage study results
from \emph{QSS} section 2.8.2
}
\details{
See \emph{QSS} Table 3.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
