% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinawomen.R
\docType{data}
\name{chinawomen}
\alias{chinawomen}
\title{Sex Ratio and the Price of Agricultural Crops in China}
\format{A data frame with 51766 rows and 9 variables:
\describe{
 \item{ admin }{ integer: unique county identifier }
 \item{ biryr }{ integer: year of cohort (birth year) }
 \item{ birpop }{ integer: birth population in a given year }
 \item{ han }{ numeric }
 \item{ sex }{ numeric: proportion of males in the birth cohort }
 \item{ teasown }{ numeric: quantity of tea sown in the county }
 \item{ orch }{ numeric: quantity of orchard-type crops planted in the county }
 \item{ cashcrop }{ numeric: quantity of cash crops planted in the county }
 \item{ post }{ integer: indicator variable for the introduction of price reforms }
}}
\usage{
chinawomen
}
\description{
Demographic and agricultural crop data for individual counties in China.
}
\details{
See \emph{QSS} Table 7.4.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { Missing women and the price of tea in China: The effect of sex-specific earnings
on sex imbalance.” \emph{Quarterly Journal of Economics}, vol. 123, no. 3, pp. 1251–1285.
}
}
}
\keyword{datasets}
