% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/congress.R
\docType{data}
\name{congress}
\alias{congress}
\title{Legislative Ideal Points Data}
\format{A data frame with 14552 rows and 7 variables:
\describe{
 \item{ congress }{ integer: congressional session number }
 \item{ district }{ integer: district number of the congressional representative }
 \item{ state }{ character: state of the congressional representative }
 \item{ party }{ character: party of the congressional representative }
 \item{ name }{ character: name of the congressional representative }
 \item{ dwnom1 }{ numeric: DW-NOMINATE score (first dimension) }
 \item{ dwnom2 }{ numeric: DW-NOMINATE score (second dimension) }
}}
\usage{
congress
}
\description{
Contains the estimated ideal points of all legislators who served in
the House of Representatives from the 80th (1947–1948) to the 112th (2011–2012) Congresses.
}
\details{
See \emph{QSS} Table 3.3.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item {Nolan McCarty, Keith T. Poole, and Howard Rosenthal (2006) \emph{Polarized America: The
Dance of Ideology and Unequal Riches. MIT Press.} doi: 10.1017/S1537592707071873 }
}
}
\keyword{datasets}
