% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constitution.R
\docType{data}
\name{constitution}
\alias{constitution}
\title{Analyzing the Preambles of Constitutions}
\format{A data frame with 155 rows and 3 variables:
\describe{
 \item{ country }{ character }  \item{ year }{ integer }  \item{ preamble }{ character } 
}}
\usage{
constitution
}
\description{
Analyzing the Preambles of Constitutions
}
\details{
See \emph{QSS} Table 5.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

