% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filedrawer.R
\docType{data}
\name{filedrawer}
\alias{filedrawer}
\title{TESS Program Academic Research Data}
\format{A data frame with 221 rows and 5 variables:
\describe{
 \item{ id }{ integer }  \item{ journal }{ character }  \item{ DV }{ character }  \item{ IV }{ character }  \item{ max.h }{ integer } 
}}
\usage{
filedrawer
}
\description{
TESS Program Academic Research Data
}
\details{
See \emph{QSS} Table 7.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

