% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLCensusVTD.R
\docType{data}
\name{FLCensusVTD}
\alias{FLCensusVTD}
\title{Florida Census Data at the Voting District Level}
\format{A data frame with 8871 rows and 8 variables:
\describe{
 \item{ county }{ integer }  \item{ VTD }{ integer }  \item{ total.pop }{ integer }  \item{ white }{ numeric }  \item{ black }{ numeric }  \item{ hispanic }{ numeric }  \item{ api }{ numeric }  \item{ others }{ numeric } 
}}
\usage{
FLCensusVTD
}
\description{
Florida Census Data at the Voting District Level
}
\details{
See \emph{QSS} Table 6.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

