% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/florida.R
\docType{data}
\name{florida}
\alias{florida}
\title{1996 and 2000 US Presidential Election Data for Florida Counties}
\format{A data frame with 67 rows and 7 variables:
\describe{
 \item{ county }{ character: county name }
 \item{ Clinton96 }{ integer: Clinton’s votes in 1996 }
 \item{ Dole96 }{ integer: Dole’s votes in 1996 }
 \item{ Perot96 }{ integer: Perot’s votes in 1996 }
 \item{ Bush00 }{ integer: Bush’s votes in 2000 }
 \item{ Gore00 }{ integer: Gore’s votes in 2000 }
 \item{ Buchanan00 }{ integer: Buchanan’s votes in 2000 }
}}
\usage{
florida
}
\description{
1996 and 2000 election outcomes in each Florida county by candidate.
}
\details{
See \emph{QSS} Table 4.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
