% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLVoters.R
\docType{data}
\name{FLVoters}
\alias{FLVoters}
\title{Florida Registered Voter List Sample Data}
\format{
A data frame with 10000 rows and 6 variables:
\describe{
 \item{ surname }{ character: surname }
 \item{ county }{ integer: county ID of the voter’s residence }
 \item{ VTD }{ integer: voting district ID of the voter’s residence }
 \item{ age }{ integer: age }
 \item{ gender }{ character: gender: m = male and f = female }
 \item{ race }{ character: self-reported race }
}
}
\usage{
FLVoters
}
\description{
Sample of 10000 from registered Florida voters.
}
\details{
See \emph{QSS} Table 6.1.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
