% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gay.R
\docType{data}
\name{gay}
\alias{gay}
\title{Gay Marriage Study Data}
\format{A data frame with 69592 rows and 4 variables:
\describe{
 \item{ study }{ integer }  \item{ treatment }{ character }  \item{ wave }{ integer }  \item{ ssm }{ integer } 
}}
\usage{
gay
}
\description{
Gay Marriage Study Data
}
\details{
See \emph{QSS} Table 2.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
