% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gayreshaped.R
\docType{data}
\name{gayreshaped}
\alias{gayreshaped}
\title{Reshaped Gay Marrige Study Data}
\format{A data frame with 11948 rows and 6 variables:
\describe{
 \item{ study }{ integer }  \item{ treatment }{ character }  \item{ therm1 }{ integer }  \item{ therm2 }{ integer }  \item{ therm3 }{ integer }  \item{ therm4 }{ integer } 
}}
\usage{
gayreshaped
}
\description{
Reshaped Gay Marrige Study Data
}
\details{
See \emph{QSS} Table 3.5.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

