% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\docType{data}
\name{names}
\alias{names}
\title{US Census Bureau Surname List Data}
\format{A data frame with 151671 rows and 7 variables:
\describe{
 \item{ surname }{ character }  \item{ count }{ integer }  \item{ pctwhite }{ numeric }  \item{ pctblack }{ numeric }  \item{ pctapi }{ numeric }  \item{ pcthispanic }{ numeric }  \item{ pctothers }{ numeric } 
}}
\usage{
names
}
\description{
US Census Bureau Surname List Data
}
\details{
See \emph{QSS} Table 6.3.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

