% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls08.R
\docType{data}
\name{polls08}
\alias{polls08}
\title{2008 US Presidential State Polls Data}
\format{A data frame with 1332 rows and 5 variables:
\describe{
 \item{ state }{ character }  \item{ Pollster }{ character }  \item{ Obama }{ integer }  \item{ McCain }{ integer }  \item{ middate }{ Date } 
}}
\usage{
polls08
}
\description{
2008 US Presidential State Polls Data
}
\details{
See \emph{QSS} Table 4.2.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
