% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres08.R
\docType{data}
\name{pres08}
\alias{pres08}
\title{2008 US Presidential Election Data}
\format{A data frame with 51 rows and 5 variables:
\describe{
 \item{ state.name }{ character: unabbreviated name of the state }
 \item{ state }{ character: abbreviated name of the state }
 \item{ Obama }{ integer: Obama’s vote share (percentage) }
 \item{ McCain }{ integer: McCain’s vote share (percentage) }
 \item{ EV }{ integer: number of Electoral College votes for the state }
}}
\usage{
pres08
}
\description{
2008 US presidential election outcomes by state.
}
\details{
See \emph{QSS} Table 4.1.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { David Rothschild (2009) “Forecasting elections: Comparing prediction markets,
polls, and their biases.” \emph{Public Opinion Quarterly}, vol. 73, no. 5, pp. 895–916.
doi = 10.1093/poq/nfp082 }
}
}
\keyword{datasets}
