% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/published.R
\docType{data}
\name{published}
\alias{published}
\title{Published Journal Article Data}
\format{
A data frame with 53 rows and 5 variables:
\describe{
 \item{ id.p }{ integer: paper identifier }
 \item{ cond.s }{ integer: number of conditions in the study }
 \item{ out.s }{ integer: number of outcome variables in the study }
 \item{ cond.p }{ integer: number of conditions presented in the paper }
 \item{ out.p }{ integer: number of outcome variables used in the paper }
}
}
\usage{
published
}
\description{
Contains information about 53 published journal articles based on TESS
(Time-Sharing Experiments in the Social Sciences) projects.
}
\details{
See \emph{QSS} Table 7.6.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { Annie Franco, Neil Malhotra, and Gabor Simonovits (2014) “Publication bias in the
social sciences: Unlocking the file drawer.” \emph{Science}, vol. 345, no. 6203, pp. 1502–1505.
doi = 10.1126/science.1255484 }
\item { Annie Franco, Neil Malhotra, and Gabor Simonovits (2015) “Underreporting in political
science survey experiments: Comparing questionnaires to published results.” \emph{Political Analysis},
vol. 23, pp. 206–312. doi = 10.1093/pan/mpv006 }
}
}
\keyword{datasets}
