% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/russia2012.R
\docType{data}
\name{russia2012}
\alias{russia2012}
\title{Russian Election Data (2012)}
\format{A data frame with 95552 rows and 3 variables:
\describe{
 \item{ N }{ integer: total number of voters in a precinct }
 \item{ turnout }{ integer: total turnout in a precinct }
 \item{ votes }{ integer: total number of votes for the winner in a precinct }
}}
\usage{
russia2012
}
\description{
Contains the election results for the 2012 Russian presidential election.
}
\details{
See \emph{QSS} Table 6.5.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { Arturas Rozenas (2016) “Inferring election fraud from distributions of voteproportions.” }
}
}
\keyword{datasets}
