% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAR.R
\docType{data}
\name{STAR}
\alias{STAR}
\title{STAR Project Data}
\format{A data frame with 6325 rows and 6 variables:
\describe{
 \item{ race }{ integer }  \item{ classtype }{ integer }  \item{ yearssmall }{ integer }  \item{ hsgrad }{ integer }  \item{ g4math }{ integer }  \item{ g4reading }{ integer } 
}}
\usage{
STAR
}
\description{
STAR Project Data
}
\details{
See \emph{QSS} Table 2.6.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

