% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sweden.R
\docType{data}
\name{Sweden}
\alias{Sweden}
\title{Fertility and Mortality estimates data (Sweden)}
\format{
A data frame with 30 rows and 8 variables:
\describe{
 \item{ country }{ character: abbreviated country name }
 \item{ period }{ characte: period during which data are collectedr }
 \item{ age }{ character: age group }
 \item{ births }{ numeric: number of births (in thousands), i.e., the number of children born to women of the age group }
 \item{ deaths }{ numeric: number of deaths (in thousands) }
 \item{ py.men }{ numeric: person-years for men (in thousands) }
 \item{ py.women }{ numeric: person-years for women (in thousands) }
 \item{ l_x }{ numeric }
}
}
\usage{
Sweden
}
\description{
Data set containing basic demographic quantities during two time periods: 1950-1955 and 2005-2010.
}
\details{
See \emph{QSS} Table 1.4.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item { United Nations, Department of Economic and Social Affairs, Population Division (2013).
\emph{World Population Prospects: The 2012 Revision, DVD Edition.} }
}
}
\keyword{datasets}
