% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade.R
\docType{data}
\name{trade}
\alias{trade}
\title{International Trade Network Data}
\format{A data frame with 114980 rows and 5 variables:
\describe{
 \item{ .row }{ integer }  \item{ country1 }{ character }  \item{ country2 }{ character }  \item{ year }{ integer }  \item{ exports }{ numeric } 
}}
\usage{
trade
}
\description{
International Trade Network Data
}
\details{
See \emph{QSS} Table 5.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

