% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\docType{data}
\name{transfer}
\alias{transfer}
\title{Brazilian Government Transfer Data}
\format{A data frame with 1787 rows and 9 variables:
\describe{
 \item{ .row }{ integer }  \item{ state }{ character }  \item{ region }{ character }  \item{ literate91 }{ numeric }  \item{ educ91 }{ numeric }  \item{ poverty91 }{ numeric }  \item{ poverty80 }{ numeric }  \item{ educ80 }{ numeric }  \item{ pop82 }{ integer } 
}}
\usage{
transfer
}
\description{
Brazilian Government Transfer Data
}
\details{
See \emph{QSS} Table 4.12.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}

