% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.following.R
\docType{data}
\name{twitter.following}
\alias{twitter.following}
\title{Senator Follow / Following Data}
\format{
A data frame with 3859 rows and 2 variables:
\describe{
 \item{ following }{ character: Twitter screen name of the following senator }
 \item{ followed }{ character: Twitter screen name of the followed senator }
}
}
\usage{
twitter.following
}
\description{
Lists pairs of the Twitter screen names of following and followed politicians.
}
\details{
See \emph{QSS} Table 5.4.
}
\references{
\itemize{
\item{ Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}. }
\item {Pablo Barberá; Birds of the Same Feather Tweet Together:
Bayesian Ideal Point Estimation Using Twitter Data. \emph{Political Analysis} 2015;
23 (1): 76-91. doi: 10.1093/pan/mpu011 }
}
}
\keyword{datasets}
