% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter.senator.R
\docType{data}
\name{twitter.senator}
\alias{twitter.senator}
\title{Senator Information Data}
\format{A data frame with 91 rows and 4 variables:
\describe{
 \item{ screen_name }{ character: Twitter screen name }
 \item{ name }{ character: name of senator }
 \item{ party }{ character: party (D = Democrat, R = Republican, I = Independent) }
 \item{ state }{ character: state abbreviation }
}}
\usage{
twitter.senator
}
\description{
Senator Information Data
}
\details{
See \emph{QSS} Table 5.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
