% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNpop.R
\docType{data}
\name{UNpop}
\alias{UNpop}
\title{World population estimates data}
\format{A data frame with 7 rows and 2 variables:
\describe{
 \item{ year }{ integer }  \item{ world.pop }{ integer } 
}}
\usage{
UNpop
}
\description{
World population estimates data
}
\details{
See \emph{QSS} Table 1.2.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
