% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignettes.R
\docType{data}
\name{vignettes}
\alias{vignettes}
\title{Political Efficacy Data}
\format{A data frame with 781 rows and 6 variables:
\describe{
 \item{ self }{ integer: self-assessment response }
 \item{ alison }{ integer: response to the Alison vignette }
 \item{ jane }{ integer: response to the Jane vignette }
 \item{ moses }{ integer: response to the Moses vignette }
 \item{ china }{ integer: 1 for China and 0 for Mexico }
 \item{ age }{ integer: age of respondent in years }
}}
\usage{
vignettes
}
\description{
Political Efficacy Data
}
\details{
See \emph{QSS} Table 3.7.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
