% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World.R
\docType{data}
\name{World}
\alias{World}
\title{Fertility and Mortality estimates data (World)}
\format{A data frame with 30 rows and 7 variables:
\describe{
 \item{ country }{ character: abbreviated country name }
 \item{ period }{ characte: period during which data are collectedr }
 \item{ age }{ character: age group }
 \item{ births }{ numeric: number of births (in thousands), i.e., the number of children born to women of the age group }
 \item{ deaths }{ numeric: number of deaths (in thousands) }
 \item{ py.men }{ numeric: person-years for men (in thousands) }
 \item{ py.women }{ numeric: person-years for women (in thousands) }
 \item{ l_x }{ numeric }#'
}}
\usage{
World
}
\description{
Fertility and Mortality estimates data (World)
}
\details{
See \emph{QSS} Table 1.4.
}
\references{
Imai, Kosuke. 2017. \emph{Quantitative Social Science: An Introduction}.
Princeton University Press. \href{http://press.princeton.edu/titles/11025.html}{URL}.
}
\keyword{datasets}
